packages/agents/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "天気データ分析AIエージェント - 共通ライブラリ"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "boto3>=1.34",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared"]
